Moves source files in src directory.
