+++ server/src/helpers/voteModelUpdates2.js
