+++ server/src/models/Poll.js
