+++ server/src/models/Vote.js
