+++ server/src/routes/api/friends.js
