+++ server/src/routes/api/vote.js
