+++ server/src/routes/ping.js
