+++ server/src/routes/utils/getVotes.js
