+++ server/src/validation/friendList.js
