+++ server/src/validation/register.js
